[rename] config 및 common을 global 패키지 아래로 이동
