+++ src/main/java/Ness/Backend/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package Ness.Backend.config;
+package Ness.Backend.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
